requirements.txt
pyzmq==25.1.2
ipykernel==6.29.5
streamlit
openai
langchain
langchain-openai
langchain-community
langchainhub
langchain-chroma
beautifulsoup4
ipykernel
python-dotenv
langgraph 
langsmith 
langchain_anthropic
tavily-python 
langchain_community
psycopg2
duckdb==1.1.3
polars==1.13.0